requirements.txt
ffmpeg-python==0.2.0
PyQt5==5.15.4